Start implementing config for the d1
