+++ platforms/allwinner-d1/Cargo.toml
@@ -1,7 +1,7 @@
 cargo-features = ["per-package-target", "profile-rustflags"]
 
 [package]
-name = "mnemos-d1-core"
+name = "mnemos-d1"
 version = "0.1.0"
 edition = "2021"
 description = """
@@ -12,22 +12,37 @@ repository = "https://github.com/tosc-rs/mnemos"
 homepage = "https://mnemos.dev"
 readme = "./README.md"
 license = "MIT OR Apache-2.0"
+forced-target = "riscv64imac-unknown-none-elf"
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[build-dependencies]
+d1-config = { path = "./d1-config" }
+serde = { version = "1.0.178", features = ["derive"] }
+mnemos-config = { path = "../../source/config", features = ["use-std"] }
+mnemos = { path = "../../source/kernel" }
+
 [dependencies]
+serde = { version = "1.0.178", features = ["derive"], default-features = false }
+d1-config = { path = "./d1-config" }
+mnemos-config = { path = "../../source/config" }
+
 d1-pac = "0.0.31"
 critical-section = "1.1.1"
 riscv-rt = "0.11.0"
 
 # kernel
 [dependencies.mnemos]
-path = "../../../source/kernel"
+path = "../../source/kernel"
 default-features = false
 
 [dependencies.tracing]
 version = "0.1.37"
 features = ["attributes"]
 default-features = false
+
+[dependencies.mnemos-beepy]
+path = "../beepy"
 
 [dependencies.embedded-graphics]
 version = "0.7.1"
