--- platforms/allwinner-d1/boards/.gitignore
@@ -1,2 +1,0 @@
-**/*.bin
-**/target
