--- platforms/allwinner-d1/boards/Cargo.toml
@@ -1,60 +1,0 @@
-cargo-features = ["per-package-target"]
-
-[package]
-name = "mnemos-d1"
-version = "0.1.0"
-edition = "2021"
-repository = "https://github.com/tosc-rs/mnemos"
-homepage = "https://mnemos.dev"
-readme = "./README.md"
-license = "MIT OR Apache-2.0"
-forced-target = "riscv64imac-unknown-none-elf"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-### bin targets for each board ###
-[lib]
-test = false
-bench = false
-
-[[bin]]
-name = "lichee-rv"
-test = false
-bench = false
-
-[[bin]]
-name = "mq-pro"
-test = false
-bench = false
-
-### deps ###
-
-[dependencies]
-# D1 core
-mnemos-d1-core = { path = "../core" }
-
-# rust-embedded crates
-riscv = "0.10"
-riscv-rt = "0.11.0"
-d1-pac = "0.0.31"
-embedded-hal = "0.2.7"
-
-[dependencies.futures]
-version = "0.3.21"
-default-features = false
-features = ["async-await"]
-
-# kernel
-[dependencies.mnemos]
-path = "../../../source/kernel"
-features = ["serial-trace"]
-
-# tracing 0.2
-[dependencies.tracing]
-git = "https://github.com/tokio-rs/tracing"
-# branch = "master"
-features = ["attributes"]
-default-features = false
-
-[dependencies.mnemos-beepy]
-path = "../../beepy"
