--- platforms/allwinner-d1/core/README.md
@@ -1,4 +1,0 @@
-# Allwinner D1 Core
-
-A hardware abstraction library for the Allwinner D1, targeted
-at use in the mnemos operating system.
