+++ platforms/allwinner-d1/d1-config/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "d1-config"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = { version = "1.0.178", features = ["derive"], default-features = false }
