+++ platforms/allwinner-d1/d1.toml
@@ -1,0 +1,22 @@
+[kernel]
+max_drivers = 16
+timer_granularity = { secs = 0, nanos = 333 }
+
+[services.spawnulator]
+enabled = true
+
+[services.keyboard_mux]
+enabled = true
+
+[services.serial_mux]
+enabled = true
+
+[services.sermux_loopback]
+enabled = true
+
+[services.sermux_hello]
+enabled = true
+
+[platform.i2c]
+enabled = true
+mapping = "LicheeRvTwi0"
