+++ platforms/allwinner-d1/memory.x
