+++ platforms/allwinner-d1/oranda.json
