+++ platforms/allwinner-d1/src/dmac/descriptor.rs
