+++ platforms/allwinner-d1/src/dmac/mod.rs
