+++ platforms/allwinner-d1/src/drivers/spim.rs
