+++ platforms/allwinner-d1/src/ram.rs
