+++ platforms/allwinner-d1/src/timer.rs
