+++ platforms/allwinner-d1/src/trap.rs
